Source work
Moved "managed" schema to "smbtrack" to not try and camp on a generic
word as our schema name.

Changed AppConfig source to use StartConfigurationSesson and
GetLatestConfiguration endpoints.  There's some confusing poll time
fields - I'm not sure why AWS is trying to ditate our polling
frequency.  For now I'm just logging that to see what it's giving us.
We may have to put a feedback interface from source to polling manager
so it can control the re-poll, but that seems super ugly.