Uncaught Json exceptions
The polling configuration manager wasn't catching JsonParseExceptions
that were being thrown by JsonComposite, and it caused the thread to
exit.  I've updated JsonComposite to wrap those in a checked exception
so we don't miss those in the future, but I've also added a catch-all
clause in PollingConfigManager to catch any other wild unchecked
exceptions.